+++ reference-app-web-ctv/copy_binaries.sh
@@ -1,0 +1,6 @@
+#!/bin/bash
+
+mkdir -p bin
+cp ../../bin/omweb-v1.js bin/omweb-v1.js
+cp ../bin/omid-session-client-v1.js bin/omid-session-client-v1.js
+cp ../bin/omid-validation-verification-script-v1.js bin/omid-validation-verification-script-v1.js